Refactored out a ReturnToBattlefieldAttachedEffect from Gift of Immortality
