+++ Mage.Sets/src/mage/cards/g/GiftOfImmortality.java
@@ -8,6 +8,7 @@ import mage.abilities.common.delayed.AtTheBeginOfNextEndStepDelayedTriggeredAbil
 import mage.abilities.effects.Effect;
 import mage.abilities.effects.OneShotEffect;
 import mage.abilities.effects.common.AttachEffect;
+import mage.abilities.effects.common.ReturnToBattlefieldAttachedEffect;
 import mage.abilities.keyword.EnchantAbility;
 import mage.cards.Card;
 import mage.cards.CardImpl;
@@ -75,57 +76,25 @@ class GiftOfImmortalityEffect extends OneShotEffect {
     public boolean apply(Game game, Ability source) {
         Permanent enchantment = (Permanent) game.getLastKnownInformation(source.getSourceId(), Zone.BATTLEFIELD);
         Player controller = game.getPlayer(source.getControllerId());
-        if (controller != null && enchantment != null && enchantment.getAttachedTo() != null) {
-            Permanent enchanted = (Permanent) game.getLastKnownInformation(enchantment.getAttachedTo(), Zone.BATTLEFIELD);
-            Card card = game.getCard(enchantment.getAttachedTo());
-            if (card != null && enchanted != null && card.getZoneChangeCounter(game) == enchanted.getZoneChangeCounter(game) + 1) {
-                controller.moveCards(card, Zone.BATTLEFIELD, source, game, false, false, true, null);
-                Permanent permanent = game.getPermanent(card.getId());
-                if (permanent != null) {
-                    //create delayed triggered ability
-                    Effect effect = new GiftOfImmortalityReturnEnchantmentEffect();
-                    effect.setTargetPointer(new FixedTarget(permanent, game));
-                    game.addDelayedTriggeredAbility(new AtTheBeginOfNextEndStepDelayedTriggeredAbility(effect), source);
-                }
-
-            }
-            return true;
+        if (controller == null || enchantment == null || enchantment.getAttachedTo() == null) {
+            return false;
+        }
+        Permanent enchanted = (Permanent) game.getLastKnownInformation(enchantment.getAttachedTo(), Zone.BATTLEFIELD);
+        Card card = game.getCard(enchantment.getAttachedTo());
+        if (card == null || enchanted == null || card.getZoneChangeCounter(game) != enchanted.getZoneChangeCounter(game) + 1) {
+            return false;
         }
 
-        return false;
-    }
-
-}
-
-class GiftOfImmortalityReturnEnchantmentEffect extends OneShotEffect {
-
-    public GiftOfImmortalityReturnEnchantmentEffect() {
-        super(Outcome.PutCardInPlay);
-        staticText = "Return {this} to the battlefield attached to that creature at the beginning of the next end step";
-    }
-
-    public GiftOfImmortalityReturnEnchantmentEffect(final GiftOfImmortalityReturnEnchantmentEffect effect) {
-        super(effect);
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        Card aura = game.getCard(source.getSourceId());
-        if (aura != null && game.getState().getZone(aura.getId()) == Zone.GRAVEYARD) {
-            Player controller = game.getPlayer(source.getControllerId());
-            Permanent creature = game.getPermanent(getTargetPointer().getFirst(game, source));
-            if (controller != null && creature != null) {
-                game.getState().setValue("attachTo:" + aura.getId(), creature);
-                controller.moveCards(aura, Zone.BATTLEFIELD, source, game);
-                return creature.addAttachment(aura.getId(), source, game);
-            }
+        controller.moveCards(card, Zone.BATTLEFIELD, source, game, false, false, true, null);
+        Permanent permanent = game.getPermanent(card.getId());
+        if (permanent == null) {
+            return false;
         }
 
-        return false;
-    }
-
-    @Override
-    public GiftOfImmortalityReturnEnchantmentEffect copy() {
-        return new GiftOfImmortalityReturnEnchantmentEffect(this);
+        // Create delayed triggered ability
+        Effect effect = new ReturnToBattlefieldAttachedEffect();
+        effect.setTargetPointer(new FixedTarget(permanent, game));
+        game.addDelayedTriggeredAbility(new AtTheBeginOfNextEndStepDelayedTriggeredAbility(effect), source);
+        return true;
     }
 }
